// frontend/src/api/generated/models/folder-position-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Placement } from './placement';
import {
    PlacementFromJSON,
    PlacementFromJSONTyped,
    PlacementToJSON,
    PlacementToJSONTyped,
} from './placement';

/**
 * 
 * @export
 * @interface FolderPositionJson
 */
export interface FolderPositionJson {
    /**
     * 
     * @type {string}
     * @memberof FolderPositionJson
     */
    anchorFolderId: string;
    /**
     * 
     * @type {Placement}
     * @memberof FolderPositionJson
     */
    placement: Placement;
}



/**
 * Check if a given object implements the FolderPositionJson interface.
 */
export function instanceOfFolderPositionJson(value: object): value is FolderPositionJson {
    if (!('anchorFolderId' in value) || value['anchorFolderId'] === undefined) return false;
    if (!('placement' in value) || value['placement'] === undefined) return false;
    return true;
}

export function FolderPositionJsonFromJSON(json: any): FolderPositionJson {
    return FolderPositionJsonFromJSONTyped(json, false);
}

export function FolderPositionJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): FolderPositionJson {
    if (json == null) {
        return json;
    }
    return {
        
        'anchorFolderId': json['anchorFolderId'],
        'placement': PlacementFromJSON(json['placement']),
    };
}

export function FolderPositionJsonToJSON(json: any): FolderPositionJson {
    return FolderPositionJsonToJSONTyped(json, false);
}

export function FolderPositionJsonToJSONTyped(value?: FolderPositionJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'anchorFolderId': value['anchorFolderId'],
        'placement': PlacementToJSON(value['placement']),
    };
}