// frontend/src/api/generated/models/placement.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Placement = {
    Before: 'BEFORE',
    After: 'AFTER'
} as const;
export type Placement = typeof Placement[keyof typeof Placement];


export function instanceOfPlacement(value: any): boolean {
    for (const key in Placement) {
        if (Object.prototype.hasOwnProperty.call(Placement, key)) {
            if (Placement[key as keyof typeof Placement] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PlacementFromJSON(json: any): Placement {
    return PlacementFromJSONTyped(json, false);
}

export function PlacementFromJSONTyped(json: any, ignoreDiscriminator: boolean): Placement {
    return json as Placement;
}

export function PlacementToJSON(value?: Placement | null): any {
    return value as any;
}

export function PlacementToJSONTyped(value: any, ignoreDiscriminator: boolean): Placement {
    return value as Placement;
}